fix46

+++ controllers/cards.js
@@ -1,5 +1,6 @@
 const Card = require('../models/card');
 const NotFoundError404 = require('../errors/notFoundError404');
+const ForbiddeError403 = require("../errors/forbiddeError403");
 
 const getCards = async (req, res, next) => {
   try {
@@ -35,14 +36,14 @@ const createCard = (req, res, next) => {
   const { name, link } = req.body;
   const owner = req.user._id;
   Card.create({ name, link, owner })
-    .then((card) => {
-      res.send(card);
-    })
+    .then((card) => res.send(card))
     .catch(next);
 };
 
-const deleteCard = (req, res, next) => {
+const deleteCard = async (req, res, next) => {
   const { cardId } = req.params;
+  const { owner } = JSON.parse(JSON.stringify(await Card.findOne({ _id: cardId })));
+  if (owner !== req.user._id) { next(new ForbiddeError403('Недостаточно прав для удаления карточки')); return; }
   Card.deleteOne({ _id: cardId }, { runValidators: true }).then((card) => {
     if (!card.deletedCount) { next(new NotFoundError404('Карточки не существует')); return; }
     res.send(card);
